move labelmanager to analyse
