+++ misinformation/utils.py
@@ -2,8 +2,6 @@ import glob
 import os
 from pandas import DataFrame
 import pooch
-import pandas as pd
-import json
 
 
 class DownloadResource:
@@ -100,61 +98,6 @@ def dump_df(mydict: dict) -> DataFrame:
     return DataFrame.from_dict(mydict)
 
 
-class LabelManager:
-    def __init__(self):
-        self.labels_code = None
-        self.labels = None
-        self.f_labels = None
-        self.f_labels_code = None
-        self.load()
-
-    def load(self):
-        self.labels_code = pd.read_excel(
-            "./misinformation/test/data/EUROPE_APRMAY20_data_variable_labels_coding.xlsx",
-            sheet_name="variable_labels_codings",
-        )
-        self.labels = pd.read_csv(
-            "./misinformation/test/data/Europe_APRMAY20data190722.csv",
-            sep=",",
-            decimal=".",
-        )
-        self.map = self.read_json("./misinformation/data/map_test_set.json")
-
-    def read_json(self, name):
-        with open("{}".format(name)) as f:
-            mydict = json.load(f)
-        return mydict
-
-    def get_orders(self):
-        return [i["order"] for i in self.map.values()]
-
-    def filter_from_order(self, orders: list):
-        cols = []
-        for order in orders:
-            col = self.labels_code.iloc[order - 1, 1]
-            cols.append(col.lower())
-
-        self.f_labels_code = self.labels_code.loc[
-            self.labels_code["order"].isin(orders)
-        ]
-        self.f_labels = self.labels[cols]
-
-    def gen_dict(self):
-        labels_dict = {}
-        if self.f_labels is None:
-            print("No filtered labels found")
-            return labels_dict
-
-        cols = self.f_labels.columns.tolist()
-        for index, row in self.f_labels.iterrows():
-            row_dict = {}
-            for col in cols:
-                row_dict[col] = row[col]
-            labels_dict[row["pic_id"]] = row_dict
-
-        return labels_dict
-
-
 if __name__ == "__main__":
     files = find_files(
         path="/home/inga/projects/misinformation-project/misinformation/data/test_no_text/"
@@ -164,10 +107,3 @@ if __name__ == "__main__":
     outdict = append_data_to_dict(mydict)
     df = dump_df(outdict)
     print(df.head(10))
-
-    # example of LabelManager for loading csv data to dict
-    lm = LabelManager()
-    orders = lm.get_orders()
-    lm.filter_from_order([1, 2, 3] + orders)
-    labels = lm.gen_dict()
-    print(labels)
